remove `ArrayDataTrait` + test
