+++ src/DataSet/ArrayDataSet.php
@@ -8,10 +8,20 @@ use Yiisoft\Validator\DataSetInterface;
 
 final class ArrayDataSet implements DataSetInterface
 {
-    use ArrayDataTrait;
+    private array $data;
 
     public function __construct(array $data = [])
     {
         $this->data = $data;
+    }
+
+    public function getAttributeValue(string $attribute): mixed
+    {
+        return $this->data[$attribute] ?? null;
+    }
+
+    public function getData(): array
+    {
+        return $this->data;
     }
 }
